Draft: [PHP8.1] Add ReturnTypeWillChange attributes where necessary. (#2763)
* [PHP8.1] Add ReturnTypeWillChange attributes where necessary.

* Add PHP8.1 in the test.yml workflow.

* Cast to integer to fix the error `Implicit conversion from float`

+++ src/Executor/Master.php
@@ -23,7 +23,7 @@ const FRAMES = ['⠋', '⠙', '⠹', '⠸', '⠼', '⠴', '⠦', '⠧', '⠇', '
 
 function spinner(string $message = ''): string
 {
-    $frame = FRAMES[((new \DateTime)->format('u') / 1e5) % count(FRAMES)];
+    $frame = FRAMES[(int) ((new \DateTime)->format('u') / 1e5) % count(FRAMES)];
     return "  $frame $message\r";
 }
